Major overhaul, see notes
Changing the library name to be inclusive of other related schemas.
Reorgananized the Character schema to provide more clarity about what the properties are for. More inline documentation would be valuable (using description keyword)
Added bootprint-json-schema to automatically build docs however a module with more options would be preferable. A script was created to generate the /docs/ folder for github.io